refactor(numerical-semigroups): split invariant execution helpers
