+++ tests/math/numerical_semigroups/test_extended.py
@@ -4,11 +4,22 @@ import pytest
 from pydantic import ValidationError
 from tests.math.numerical_semigroups._support import numerical_semigroup_error
 
+from jacobian.math.numerical_semigroups._element_invariants import (
+    compute_element_catenary_degree,
+    compute_element_delta_set,
+    compute_element_elasticity,
+)
 from jacobian.math.numerical_semigroups._factorization_operations import (
     compute_factorization_distance,
     compute_factorization_graph,
     compute_factorization_lengths,
     compute_factorizations,
+)
+from jacobian.math.numerical_semigroups._global_invariants import (
+    compute_betti_elements,
+    compute_catenary_degree,
+    compute_delta_set,
+    compute_elasticity,
 )
 from jacobian.math.numerical_semigroups._models import (
     MAX_GENERATOR,
@@ -36,15 +47,6 @@ from jacobian.math.numerical_semigroups._models import (
     MinimalPresentationResult,
     PresentationBinomialsRequest,
     PresentationBinomialsResult,
-)
-from jacobian.math.numerical_semigroups._operations import (
-    compute_betti_elements,
-    compute_catenary_degree,
-    compute_delta_set,
-    compute_elasticity,
-    compute_element_catenary_degree,
-    compute_element_delta_set,
-    compute_element_elasticity,
 )
 from jacobian.math.numerical_semigroups._presentation_operations import (
     compute_minimal_presentation,
